requirements.txt
fastapi
google-cloud-pubsub
google-cloud-storage
pyjwt
pymongo
python-multipart
requests
uvicorn[standard]
